+++ Dice/ChronosDie.cs
@@ -75,6 +75,7 @@ namespace MysteryDice.Dice
             if (randomEffect == null) return;
 
             PlaySoundBasedOnEffect(randomEffect.Outcome);
+            MysteryDice.CustomLogger.LogDebug("Rolling Effect: "+ randomEffect.Name);
             randomEffect.Use();