+++ Dice/RustyDie.cs
@@ -8,7 +8,7 @@ namespace MysteryDice.Dice
         public override void SetupRollToEffectMapping()
         {
             RollToEffect.Add(1, new EffectType[] { EffectType.Awful });
-            RollToEffect.Add(2, new EffectType[] { EffectType.Bad });
+            RollToEffect.Add(2, new EffectType[] { EffectType.Bad, EffectType.Mixed });
             RollToEffect.Add(3, new EffectType[] { });
             RollToEffect.Add(4, new EffectType[] { });
             RollToEffect.Add(5, new EffectType[] { });
@@ -22,19 +22,22 @@ namespace MysteryDice.Dice
 
             IEffect randomEffect = GetRandomEffect(diceRoll, Effects);
 
-            if(randomEffect == null)
+            if(randomEffect == null) 
             {
                 switch (diceRoll)
                 {
                     case 3:
+                        MysteryDice.CustomLogger.LogDebug("Rolling Effect: Rusty 3");
                         Networker.Instance.JackpotServerRPC(PlayerUser.actualClientId, UnityEngine.Random.Range(1, 2));
                         Misc.SafeTipMessage($"Rolled 3", "Spawning some scrap");
                         break;
                     case 4:
+                        MysteryDice.CustomLogger.LogDebug("Rolling Effect: Rusty 4");
                         Networker.Instance.JackpotServerRPC(PlayerUser.actualClientId, UnityEngine.Random.Range(3, 4));
                         Misc.SafeTipMessage($"Rolled 4", "Spawning scrap");
                         break;
                     case 5:
+                        MysteryDice.CustomLogger.LogDebug("Rolling Effect: Rusty 5");
                         Networker.Instance.JackpotServerRPC(PlayerUser.actualClientId, UnityEngine.Random.Range(5, 6));
                         Misc.SafeTipMessage($"Rolled 5", "Spawning more scrap");
                         break;
@@ -49,19 +52,23 @@ namespace MysteryDice.Dice
                         switch (diceRoll2)
                         {
                             case 1:
+                                MysteryDice.CustomLogger.LogDebug("Rolling Effect: Rusty 6 Eggs");
                                 Networker.Instance.SameScrapServerRPC(PlayerUser.actualClientId, UnityEngine.Random.Range(4, 6), "Easter egg");
                                 Misc.SafeTipMessage($"Hop Hop", "Explosive Eggs?");
                                 break;
                             case 2:
+                                MysteryDice.CustomLogger.LogDebug("Rolling Effect: Rusty 6 Christmas");
                                 Networker.Instance.SameScrapServerRPC(PlayerUser.actualClientId, UnityEngine.Random.Range(4, 7),"Gift");
                                 Misc.SafeTipMessage($"HO HO HO", "Christmas Time!");
                                 break;
                             case 3:
+                                MysteryDice.CustomLogger.LogDebug("Rolling Effect: Rusty 6 SSDD");
                                 var item = RM.currentLevel.spawnableScrap[RM.GetRandomWeightedIndex(weights)].spawnableItem;
                                 Networker.Instance.SameScrapServerRPC(PlayerUser.actualClientId, UnityEngine.Random.Range(6, 8),item.itemName);
                                 Misc.SafeTipMessage($"Wat?", "It's all the same?!?");
                                 break;
                             default:
+                                MysteryDice.CustomLogger.LogDebug("Rolling Effect: Rusty 6 Normal");
                                 Networker.Instance.JackpotServerRPC(PlayerUser.actualClientId, UnityEngine.Random.Range(7, 8));
                                 Misc.SafeTipMessage($"Rolled 6", "Spawning a lot of scrap!");
                                 break;
@@ -73,6 +80,7 @@ namespace MysteryDice.Dice
             else
             {
                 PlaySoundBasedOnEffect(randomEffect.Outcome);
+                MysteryDice.CustomLogger.LogDebug("Rolling Effect: "+ randomEffect.Name);
                 randomEffect.Use();
                 
                 var who = !wasEnemy ? PlayerUser.playerUsername : "An Enemy";